pyproject.toml
[project]
name = "minicurso-ia-unisagrado"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "easyocr>=1.7.2",
    "google-genai>=1.50.0",
    "pillow>=12.0.0",
    "selenium>=4.38.0",
    "webdriver-manager>=4.0.2",
]
